Reuse objects in object pool
Because allocating objects often and garbage collecting is expensive,
instead of creating and releasing objects to the list of objects, object
pool is used.

Objects in the pool will be marked as either used or unused.
`iterator` will iterate only through the used objects.

To avoid allocating and releasing iterators on every frame, one iterator
will be reused per object pool.