Add files via upload
Since the output of scenes in the "render()" function has been converted into a loop, it is necessary to follow the sequence of adding scenes to the array. For example, the "Eyes" scene should be after "Background". and then the "Butterflies" and others. Otherwise, if it happens that, for example, "butterflies" and "eyes" will be played at the same time, then "Eyes" will be drawn over "Butterflies", which should not be the case.
Similarly with "CristalBlick" - it should be higher than the "Eyes".
The alphabetical sequence will not work.